# Query for item level queue length conditionally