## What this does
TLDR: Checks the `queueTotalLength` from the instance first. Only query `queueTotalLength` at the item level if that is non-zero.

* Removes `queueTotalLength` from `item_fields`. Adds it back explicitly to the `boundWithItem` portion of the instance query. I don't think we can make assumptions about bound withs here.
* Adds `queueTotalLength` to the instance level of the instance query
* If `queueTotalLength` is 0 for the instance, sets all the item level `queueTotalLength` to 0. Based on the [API docs](https://s3.amazonaws.com/foliodocs/api/mod-circulation/r/circulation.html) and what I'm seeing I think this a valid assumption (e.g., `/circulation/requests/queue/instance/#{id}` contains all the requests for items in that instance)
* If `queueTotalLength` is non-zero, run a specific query for the item level queue lengths and merge that into the instance data

## Why would we want to do this?
Having `queueTotalLength` in `item_fields` causes folio-graphql to run one unique Folio API query (`/circulation/requests/queue/item/#{id}`) for each item which can be time consuming. We have instances with many items. Most instances/items do not have queues.

## How much time does this save when the instance queue length is 0?
Some. Not enough, but meaningful for instances with a large number of items.

Before:
```ruby
sul-requests(dev)*   puts Benchmark.measure { FolioGraphqlClient.new.instance(hrid: 'L17') }
sul-requests(dev)> end
  0.006563   0.004457   0.011020 (  7.055010)
  0.005537   0.003091   0.008628 (  6.775460)
  0.004782   0.001473   0.006255 (  6.520238)
sul-requests(dev)* 3.times do
sul-requests(dev)*   puts Benchmark.measure { FolioGraphqlClient.new.instance(hrid: 'L284') }
sul-requests(dev)> end
  0.047194   0.017478   0.064672 (  9.883716)
  0.006027   0.002250   0.008277 (  9.527961)
  0.005904   0.002884   0.008788 (  9.650898)
sul-requests(dev)* 3.times do
```

After:
```ruby
sul-requests(dev)* 3.times do
sul-requests(dev)*   puts Benchmark.measure { FolioGraphqlClient.new.instance(hrid: 'L17') }
sul-requests(dev)> end
  0.009764   0.005406   0.015170 (  6.293825)
  0.008824   0.001706   0.010530 (  5.968284)
  0.005665   0.002254   0.007919 (  5.830505)
sul-requests(dev)* 3.times do
sul-requests(dev)*   puts Benchmark.measure { FolioGraphqlClient.new.instance(hrid: 'L284') }
sul-requests(dev)> end
  0.008519   0.003153   0.011672 (  8.610549)
  0.010763   0.011110   0.021873 (  8.345573)
  0.008871   0.001811   0.010682 (  8.181223)
```